xl: refactor functions to xl-v1-common.go xl-v1-utils.go. (#1357)
