+++ xl-v1-readfile.go
@@ -21,95 +21,7 @@ import (
 	"fmt"
 	"io"
 	slashpath "path"
-	"strconv"
 )
-
-// checkBlockSize return the size of a single block.
-// The first non-zero size is returned,
-// or 0 if all blocks are size 0.
-func checkBlockSize(blocks [][]byte) int {
-	for _, block := range blocks {
-		if len(block) != 0 {
-			return len(block)
-		}
-	}
-	return 0
-}
-
-// calculate the blockSize based on input length and total number of
-// data blocks.
-func getEncodedBlockLen(inputLen, dataBlocks int) (curBlockSize int) {
-	curBlockSize = (inputLen + dataBlocks - 1) / dataBlocks
-	return
-}
-
-// Returns slice of disks needed for ReadFile operation:
-// - slice returing readable disks.
-// - fileMetadata
-// - bool value indicating if selfHeal is needed.
-// - error if any.
-func (xl XL) getReadableDisks(volume, path string) ([]StorageAPI, fileMetadata, bool, error) {
-	partsMetadata, errs := xl.getPartsMetadata(volume, path)
-	highestVersion := int64(0)
-	versions := make([]int64, len(xl.storageDisks))
-	quorumDisks := make([]StorageAPI, len(xl.storageDisks))
-	notFoundCount := 0
-	// If quorum says errFileNotFound return errFileNotFound
-	for _, err := range errs {
-		if err == errFileNotFound {
-			notFoundCount++
-		}
-	}
-	if notFoundCount > xl.readQuorum {
-		return nil, fileMetadata{}, false, errFileNotFound
-	}
-	for index, metadata := range partsMetadata {
-		if errs[index] == nil {
-			if version := metadata.Get("file.version"); version != nil {
-				// Convert string to integer.
-				version, err := strconv.ParseInt(version[0], 10, 64)
-				if err != nil {
-					// Unexpected, return error.
-					return nil, fileMetadata{}, false, err
-				}
-				if version > highestVersion {
-					highestVersion = version
-				}
-				versions[index] = version
-			} else {
-				versions[index] = 0
-			}
-		} else {
-			versions[index] = -1
-		}
-	}
-	quorumCount := 0
-	for index, version := range versions {
-		if version == highestVersion {
-			quorumDisks[index] = xl.storageDisks[index]
-			quorumCount++
-		} else {
-			quorumDisks[index] = nil
-		}
-	}
-	if quorumCount < xl.readQuorum {
-		return nil, fileMetadata{}, false, errReadQuorum
-	}
-	var metadata fileMetadata
-	for index, disk := range quorumDisks {
-		if disk == nil {
-			continue
-		}
-		metadata = partsMetadata[index]
-		break
-	}
-	// FIXME: take care of the situation when a disk has failed and been removed
-	// by looking at the error returned from the fs layer. fs-layer will have
-	// to return an error indicating that the disk is not available and should be
-	// different from ErrNotExist.
-	doSelfHeal := quorumCount != len(xl.storageDisks)
-	return quorumDisks, metadata, doSelfHeal, nil
-}
 
 // ReadFile - read file
 func (xl XL) ReadFile(volume, path string, offset int64) (io.ReadCloser, error) {
